Added first usable revision of self-photo avatar capture.
This unfortunately requires the app to allow portrait orientation.
Avatar images are arbitrarily 90*60 for now. Additionally I pulled in
Trevor Harmon's resize code, as I'm too tired to be sure a solution of
my own would work anywhere near as well.. did have to hack that
slightly to avoid an annoying problem in iOS 6 - apparently something
that was "fixed" in 5 has now been SUPER-fixed to behave as it did in
4, so there's special 5-only image rotation cases.